requirements.txt
pyTelegramBotAPI==4.8.0
python-dotenv==0.19.2
SQLAlchemy==1.4.44
sqlalchemy-utils==0.38.3
psycopg2-binary==2.9.5